Important changes: refactored the dynamo.f90. The main objectives were: the general improvement of clarity and to allow bad snapshots to be ignored. WARNING: code curretly BROKEN.
Previously, when the equations misbehaved and a unphysical value for B was found, the calculation restarted for that galaxy. But we do not want to restart the galaxy in this case: we want to reduce the timestep and rerun that particular snapshot, without erasing previous work on other snapshots. If it fails many times, we want the code to ignore that particular snapshot and proceed to the next -- using a new seed as initial condition. The bogus snapshot should be tagged in the output with invalid entries -- using e.g. negative values -- allowing us to either ignore or check them.

At the moment, I am fighting an arithmetic exception that appears after the number os steps is increased.