Sets up a floor for the density.
Several adjustments to the pressure calculation were made: there is now a more sensible estimate of the interval where the roots are likely to be found, with the option of such interval to be enlarge in the case of failure. If no solution was found, falls back to the value of the scale-height at the previous radius (keeping things more or less sensible, despite the error) and flags the galaxy.